+++ aofl-js-packages/polyfill-service/src/polyfill/polyfills.sample.js
@@ -2,5 +2,11 @@ export default {
   'fetch': () => import('isomorphic-fetch'),
   'Reflect': () => import('harmony-reflect'),
   'Proxy': () => import('harmony-reflect'),
-  'Array.prototype.find': () => import('array.prototype.find')
+  'Array.prototype.find': () => import('array.prototype.find'),
+  'html-imports': {
+    test() {
+      return !('import' in document.createElement('link'));
+    },
+    load: () => import('@webcomponents/webcomponentsjs/webcomponents-bundle')
+  }
 };